Tevin branch

+++ FRS/src/main/java/com/group/repository/Passenger_ScheduleRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.group.model.Passenger_Schedule;
 
-public interface Passenger_ScheduleRepository extends CrudRepository<Passenger_Schedule, Long> {
-
+public interface Passenger_ScheduleRepository extends CrudRepository<Passenger_Schedule, Integer> {
+//
 }
